+++ lib/src/models/backup.dart
@@ -5,21 +5,27 @@ part 'backup.g.dart';
 @JsonSerializable()
 class Backup {
   String uuid;
+  bool isSuccessful;
+  bool isLocked;
   String name;
-  List<dynamic>? ignoredFiles;
-  String sha256Hash;
+  List<String>? ignoredFiles;
+  String checksum;
   int bytes;
   DateTime createdAt;
-  DateTime completedAt;
+  DateTime? completedAt;
+
   Backup({
     required this.uuid,
+    required this.isSuccessful,
+    required this.isLocked,
     required this.name,
     this.ignoredFiles,
-    required this.sha256Hash,
+    required this.checksum,
     required this.bytes,
     required this.createdAt,
-    required this.completedAt,
+    this.completedAt,
   });
+
   factory Backup.fromJson(Map<String, dynamic> json) => _$BackupFromJson(json);
   Map<String, dynamic> toJson() => _$BackupToJson(this);
 }